Add files via upload
HALLOWEEN CANDY PROJECT (PYTHON)
Overview
Data cleaning and data wrangling from a survey of Halloween candy to prepare it for a machine learning project.

Data Set
The data set that we will be using is the 2017 Halloween Candy Hierarchy data set as discussed in this BoingBoing article. You can also read more about the data in the Science Creative Quarterly.

The following are the rating instructions from the survey:

Consider that feeling you get when you receive this item in your Halloween haul. Does it make you happy (JOY)? Or is it something that you automatically place in the junk pile (DESPAIR)? MEH for indifference, and you can leave it blank if you have no idea what the item is.

This data is a great example of a messy data set, especially since they allowed respondents to enter text for a number of the fields. Also, note that some of the comments in the file might be considered inappropriate to some readers but cleaning this type of data is normal in a lot of data science projects.

Our End Goal
Our end goal for this project is to clean the data so that we can then create a machine-learning model. We want to see if we can predict a person's gender based purely on their candy preferences.